+++ android/app/src/main/kotlin/com/juri117/ink2brain/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.juri117.ink2brain
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}